pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "j1"
version = "0.1.0"
description = "A Python project"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "dask[distributed]>=2023.0.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "beautifulsoup4>=4.12.0",
    "tenacity>=8.0.0",
    "lxml>=4.9.0",
    "pyarrow>=14.0.0",
    "tqdm>=4.65.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "openai>=1.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["j1"]
